Geo utils tag (#36)
This was frustrating to say the least. Basically we don't need the GEOs buildpack, the environment variable BUILD_WITH_GEOS actually takes care of that but our GEO buildpack installed after herokus environment packages which used to be fine but now breaks the build due to a security update on the heroku stack.

Also added auto db-migrations on release, gunicorn was also needed in requirements.