Areas belong to offices (#50)
* allow more than one shapefile per download link (rare edge case). will tie all the areas in the shapefile to the one office

* removing and ignoring import artifacts

* removing and ignoring import artifacts

* removing and ignoring import artifacts

* areas belong to offices, not districts

My prior data model was incorrect.

Recall that a district is a broad jurisdiction, e.g. "San Francisco",
"alameda county", etc.

And recall that an "Area" is a polygon on a map. Areas are what
shapefiles are turned into in our database.

I was incorreclty tieing the areas in a shapefile to a _district_,
instead of to individual _offices_. This was just flatout wrong.

This commit takes the prior DB relationship, which was:
districts -< offices
          -< area(s)

and moves them to:
districts -< offices -< area(s)

I'm using "-<" to denote "has many", i.e. there's a many-to-one
relationship between offices and districts.